# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job-Search Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/skills": {
            "get": {
                "tags": [
                    "Skills"
                ],
                "summary": "Get list of skills",
                "description": "Returns list of skills",
                "operationId": "get skills list",
                "responses": {
                    "200": {
                        "description": "All skills"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/skills/create": {
            "post": {
                "tags": [
                    "Skills"
                ],
                "summary": "Create new skill",
                "description": "Add new skill",
                "operationId": "create new skill",
                "requestBody": {
                    "description": "Pass skill credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "years_of_experience",
                                    "justification"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "skill"
                                    },
                                    "years_of_experience": {
                                        "type": "string",
                                        "format": "years_of_experience",
                                        "example": "1-3 years"
                                    },
                                    "justification": {
                                        "type": "string",
                                        "format": "justification",
                                        "example": "test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "skill created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/skills/update": {
            "post": {
                "tags": [
                    "Skills"
                ],
                "summary": "Update skill",
                "description": "Edit skill",
                "operationId": "update skill",
                "requestBody": {
                    "description": "Pass skill credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "skill_id",
                                    "name",
                                    "years_of_experinces",
                                    "justification"
                                ],
                                "properties": {
                                    "skill_id": {
                                        "type": "integer",
                                        "format": "skill_id",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "skill"
                                    },
                                    "years_of_experience": {
                                        "type": "string",
                                        "format": "years_of_experience",
                                        "example": "1-3 years"
                                    },
                                    "justification": {
                                        "type": "string",
                                        "format": "justification",
                                        "example": "test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "skill updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/skills/delete": {
            "post": {
                "tags": [
                    "Skills"
                ],
                "summary": "Soft delete skill",
                "description": "Soft delete skill",
                "operationId": "delete specific skill",
                "requestBody": {
                    "description": "Pass skill credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "skill_id"
                                ],
                                "properties": {
                                    "skill_id": {
                                        "type": "integer",
                                        "format": "skill_id",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "skill deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}